+++ AzureDevopsBacklog/Application/Models/SprintWorkItemDetail.cs
@@ -1,0 +1,11 @@
+﻿namespace AzureDevopsBacklog.Application.Models
+{
+    public class SprintWorkItemDetail
+    {
+        public string Title { get; set; }
+        public string Tag { get; set; }
+        public string State { get; set; }
+        public string TShirtSize { get; set; }
+        public bool IsHold { get; set; }
+    }
+}
